requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
pymupdf>=1.23.0
PyJWT>=2.0.0
werkzeug>=2.3.0
openpyxl>=3.1.0
xlrd>=2.0.0
